Small story fixes, move `FilterLogic` enum
