+++ src/components/LogWindow/index.tsx
@@ -5,8 +5,8 @@ import LogPane from "components/LogPane";
 import { RowRenderer, cache } from "components/LogRow/RowRenderer";
 import SideBar from "components/SideBar";
 import SubHeader from "components/SubHeader";
-import { LogTypes } from "constants/enums";
-import { FilterLogic, QueryParams } from "constants/queryParams";
+import { FilterLogic, LogTypes } from "constants/enums";
+import { QueryParams } from "constants/queryParams";
 import { useLogContext } from "context/LogContext";
 import { useQueryParam } from "hooks/useQueryParam";
 import { findLineIndex } from "utils/findLineIndex";
