+++ src/context/LogContext/index.tsx
@@ -1,7 +1,7 @@
 import { createContext, useCallback, useContext, useMemo, useRef } from "react";
 import { List } from "react-virtualized";
-import { LogTypes } from "constants/enums";
-import { FilterLogic, QueryParams } from "constants/queryParams";
+import { FilterLogic, LogTypes } from "constants/enums";
+import { QueryParams } from "constants/queryParams";
 import { useQueryParam } from "hooks/useQueryParam";
 import { ProcessedLogLines } from "types/logs";
 import { filterLogs } from "utils/filter";
